- Feature: make optional the junitReport directory
