+++ lib/docs.js
@@ -13,6 +13,7 @@ var params = require('./params'),
 var docs = {
 
     cmdOptions : {
+        "junitReport" : "write junit report at the end",
         "initShot": "from shot : "+config.cmd+" -i <shotname>",
         "endShot": "to shot : "+config.cmd+" -e <shotname>",
         "help" : "shows detailed info for command, <name> of straw or -s <shot> name shot : "+config.cmd+" -h <strawname>",
