+++ lib/params.js
@@ -2,6 +2,7 @@ var nopt = require("nopt");
 
 var knownOpts = {
         "help" : Boolean,
+        "junitReport" : Boolean,
         "shot" : Boolean,
         "commands" : ["all","straws","shots"],
         "list" : ["all","recipes","straws","shots","repoTypes"],
@@ -11,6 +12,7 @@ var knownOpts = {
         "level" : [ "debug", "verbose", "warning" ]
     },
     shortHands = {
+        "u": ["--junitReport"],
         "i": ["--initShot"],
         "e": ["--endShot"],
         "h" : ["--help"],
