+++ lib/shot.js
@@ -6,6 +6,7 @@ var logger = require("./logger"),
     runSequence = require('run-sequence'),
     async = require('async'),
     inquirer = require('inquirer'),
+    //Mocha = require('mocha'),
     spawn = require('child_process').spawn,
     spawnSync = require('child_process').spawnSync;
 
@@ -20,6 +21,20 @@ var Shot = function(runner){
     this.logger = logger;
     return this;
 };
+
+/*  For the mocha execution
+var mocha = new Mocha({
+    reporter: 'mocha-junit-reporter',
+    reporterOptions: {
+        mochaFile: path.join(config.junitDir,"pisco-unit.xml")
+    }
+});
+mocha.files = [path.join(config.modulesDir.module,"lib","tests","execution.js")];
+var runner = mocha.run();
+runner.on("end",function(){
+    cb(runner.failures);
+});
+*/
 
 Shot.prototype.do = function(stage){
     var operation = this.runner[stage];
