+++ lib/sour.js
@@ -7,7 +7,6 @@ var docs = require("./docs"),
     config = require("./config"),
     shooter = require("./shooter"),
     params = require("./params"),
-    Mocha = require('mocha'),
     sipper = require("./sipper"),
     logger = require("./logger");
 
@@ -28,38 +27,28 @@ var Sour = function(){
             logger.trace("#green","sour:gush","params:",params.argv.remain);
 
             var xresolve = function(){
-                /*mochaInterceptor(function(failures){
-                    if (failures===0) {
-                        console.log("-------OK--------");*/
-                        resolve.apply(this, arguments);
-                    /*}else {
-                        console.log("-------ERROR--------");
-                        reject.apply(this, arguments);
-                    }
-                });*/
+                var isOk = finalCheck();
+                if (isOk)
+                    resolve.apply(this, arguments);
+                else
+                    reject.apply(this, arguments);
             };
 
             var xreject = function(){
-                //mochaInterceptor(function(){
-                  //  console.log("-------ERROR XREJECT--------");
-                    reject.apply(this,arguments);
-                //});
+                finalCheck();
+                reject.apply(this,arguments);
             };
 
-            var mochaInterceptor = function(cb){
-                logger.info("#magenta","Write results in junit format at", path.join(config.junitDir,"pisco-unit.xml"));
-                var mocha = new Mocha({
-                    reporter: 'mocha-junit-reporter',
-                    reporterOptions: {
-                        mochaFile: path.join(config.junitDir,"pisco-unit.xml")
-                    }
-                });
-                mocha.files = [path.join(config.modulesDir.module,"lib","tests","execution.js")];
-                var runner = mocha.run();
-                runner.on("end",function(){
-                   cb(runner.failures);
-                });
+            var finalCheck = function(){
+                logger.info("#magenta","finalCheck...");
+                if (params.junitReport)
+                    writeJunitXml();
+                return true;
             };
+
+            var writeJunitXml = function(){
+                logger.info("#magenta","Write results in junit format at", path.join(config.junitDir,"pisco-unit.xml"));
+            }
 
             try {
                 var normal = check(reject);
